Day 9 task

+++ server/src/main/java/com/edutech/progressive/service/impl/CourseServiceImplJpa.java
@@ -7,6 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.edutech.progressive.entity.Course;
+import com.edutech.progressive.exception.CourseAlreadyExistsException;
+import com.edutech.progressive.exception.CourseNotFoundException;
 import com.edutech.progressive.repository.CourseRepository;
 import com.edutech.progressive.service.CourseService;
 
@@ -27,22 +29,33 @@ public class CourseServiceImplJpa implements CourseService {
 
     @Override
     public Course getCourseById(int courseId) throws Exception {
-        return courseRepository.findByCourseId(courseId);
+        return courseRepository.findById(courseId)
+                .orElseThrow(() -> new CourseNotFoundException("Course with ID " + courseId + " not found for deletion"));
     }
 
     @Override
     public Integer addCourse(Course course) throws Exception {
-        courseRepository.save(course);
-        return course.getCourseId();
+        Course existingCourse = courseRepository.findByCourseName(course.getCourseName());
+        if (existingCourse != null) {
+            throw new CourseAlreadyExistsException("Course with this name already exists, Course Name: " + course.getCourseName());
+        }
+        return courseRepository.save(course).getCourseId();
     }
 
     @Override
     public void updateCourse(Course course) throws Exception {
+        Course existingCourse = courseRepository.findByCourseName(course.getCourseName());
+        if (existingCourse != null && existingCourse.getCourseId() != course.getCourseId()) {
+            throw new CourseAlreadyExistsException("Course with this name already exists, Course Name: " + course.getCourseName());
+        }
         courseRepository.save(course);
     }
 
     @Override
     public void deleteCourse(int courseId) throws Exception {
+        if (!courseRepository.existsById(courseId)) {
+            throw new CourseNotFoundException("Course with ID " + courseId + " not found for deletion");
+        }
         courseRepository.deleteById(courseId);
     }
